@@ -1,28 +1,62 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import IconButton from 'material-ui/IconButton';
-import Snackbar from 'material-ui/Snackbar';
-import Paper from 'material-ui/Paper';
-import { Table, TableBody, TableHeader, TableHeaderColumn, TableRow, TableRowColumn } from 'material-ui/Table';
-import AddIcon from 'material-ui/svg-icons/content/add';
-import RemoveIcon from 'material-ui/svg-icons/content/clear';
-import Dialog from 'material-ui/Dialog';
-import FlatButton from 'material-ui/FlatButton';
-
+import {
+  Button, IconButton, Snackbar, Typography, CircularProgress, Dialog,
+  Tooltip, Table, TableHead, TableBody, TableRow, TableCell,
+  DialogTitle, DialogContent, DialogContentText, DialogActions,
+} from '@material-ui/core';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+import AddIcon from '@material-ui/icons/Add';
+import AttachIcon from '@material-ui/icons/CallMerge';
+import RemoveIcon from '@material-ui/icons/Clear';
 import api from '../../services/api';
 import NewAddon from './NewAddon';
+import AttachAddon from './AttachAddon';
 import ConfirmationModal from '../ConfirmationModal';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+// fastest way to check for an empty object (https://stackoverflow.com/questions/679915)
+function isEmpty(obj) {
+  let empty = true;
+  Object.keys(obj).forEach((key) => {
+    if (Object.prototype.hasOwnProperty.call(obj, key)) { empty = false; }
+  });
+  return empty;
+}
+
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiPaper: {
+      root: {
+        boxShadow: '0 !important',
+      },
+    },
+    MuiDialog: {
+      paper: {
+        width: '40%',
+      },
+    },
+  },
 });
 
 const style = {
+  iconButton: {
+    color: 'black',
+  },
+  addonDialog: {
+    width: '35%',
+  },
   tableRow: {
     height: '58px',
+  },
+  tableRowPointer: {
+    height: '58px',
+    cursor: 'pointer',
   },
   tableRowColumn: {
     title: {
@@ -60,46 +94,173 @@ export default class Addons extends Component {
     this.state = {
       addons: [],
       addon: null,
+      attachment: null,
       loading: true,
       open: false,
-      confirmOpen: false,
+      confirmAddonOpen: false,
+      confirmAttachmentOpen: false,
       message: '',
       new: false,
       submitFail: false,
       submitMessage: '',
+      attach: false,
+      addonAttachments: [],
+      addonsLoaded: false,
+      attachmentsLoaded: false,
+      currentAddon: {},
+      addonDialogOpen: false,
     };
+    this.loadAddons();
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (nextProps.active) {
-      api.getAppAddons(this.props.app).then((response) => {
-        this.setState({
-          addons: response.data,
-          loading: false,
-        });
-      });
-    }
+  componentDidMount() {
+    this._isMounted = true;
+  }
+  componentWillUnmount() {
+    this._isMounted = false;
   }
 
   getAddons() {
     return this.state.addons.map(addon => (
-      <TableRow className={addon.addon_service.name} key={addon.id} style={style.tableRow}>
-        <TableRowColumn>
+      <TableRow
+        hover
+        className={addon.addon_service.name}
+        key={addon.id}
+        style={style.tableRowPointer}
+      >
+        <TableCell
+          onClick={() => this.setState({ currentAddon: addon, addonDialogOpen: true })}
+        >
           <div style={style.tableRowColumn.title}>{addon.addon_service.name}</div>
           <div style={style.tableRowColumn.sub}>{addon.id}</div>
-        </TableRowColumn>
-        <TableRowColumn>
+        </TableCell>
+        <TableCell
+          onClick={() => this.setState({ currentAddon: addon, addonDialogOpen: true })}
+        >
           <div style={style.tableRowColumn.title}>{addon.plan.name}</div>
-        </TableRowColumn>
-        <TableRowColumn style={style.tableRowColumn.icon}>
+        </TableCell>
+        <TableCell style={style.tableRowColumn.icon}>
           <div style={style.tableRowColumn.end}>
-            <IconButton className="remove" onTouchTap={() => this.handleConfirmation(addon)}>
+            <IconButton style={style.iconButton} className="addon-remove" onClick={() => this.handleAddonConfirmation(addon)}>
               <RemoveIcon />
             </IconButton>
           </div>
-        </TableRowColumn>
+        </TableCell>
       </TableRow>
     ));
+  }
+
+  getAddonAttachments() {
+    return this.state.addonAttachments.map((attachment, index) => (
+      <TableRow
+        hover
+        className={`${attachment.name} addon-attachment-list-${index}`}
+        key={attachment.id}
+        style={style.tableRowPointer}
+      >
+        <TableCell
+          onClick={() => this.setState({ currentAddon: attachment, addonDialogOpen: true })}
+        >
+          <div style={style.tableRowColumn.title}>{attachment.name}</div>
+          <div style={style.tableRowColumn.sub}>{attachment.id}</div>
+        </TableCell>
+        <TableCell
+          onClick={() => this.setState({ currentAddon: attachment, addonDialogOpen: true })}
+        >
+          <div style={style.tableRowColumn.title}>{attachment.addon.plan.name}</div>
+        </TableCell>
+        <TableCell
+          onClick={() => this.setState({ currentAddon: attachment, addonDialogOpen: true })}
+        >
+          <div style={style.tableRowColumn.title}>{attachment.addon.app.name}</div>
+        </TableCell>
+        <TableCell style={style.tableRowColumn.icon}>
+          <div style={style.tableRowColumn.end}>
+            <IconButton style={style.iconButton} className="attachment-remove" onClick={() => this.handleAddonAttachmentConfirmation(attachment)}>
+              <RemoveIcon />
+            </IconButton>
+          </div>
+        </TableCell>
+      </TableRow>
+    ));
+  }
+
+  getDialogTitle() {
+    if (this.state.addonsLoaded && this.state.currentAddon) {
+      const currentAddon = this.state.currentAddon;
+      return (
+        <div>
+          <Typography variant="h5">Attached Apps</Typography>
+          <Typography className="addon-name" variant="subtitle1" style={{ marginTop: '10px' }}>
+            {currentAddon.addon_service ? (currentAddon.addon_service.name) : (currentAddon.name)}
+            {currentAddon.addon_service ? (` (${currentAddon.name})`) : ''}
+          </Typography>
+        </div>
+      );
+    }
+    return '';
+  }
+
+  getAppsAttachedToAddon() {
+    const addons = this.state.addons;
+    addons.forEach((addon, index) => {
+      api.getAppsAttachedToAddon(this.props.app, addon.id).then((res) => {
+        addons[index].attached_to = res.data.attached_to;
+        if (addons.every(a => (a.attached_to))) {
+          if (this._isMounted) {
+            this.setState({ addons, addonsLoaded: true });
+          }
+        }
+      });
+    });
+
+    const addonAttachments = this.state.addonAttachments;
+    addonAttachments.forEach((attachment, index) => {
+      api.getAppsAttachedToAddon(this.props.app, attachment.addon.id).then((res) => {
+        addonAttachments[index].attached_to = res.data.attached_to;
+        if (addonAttachments.every(a => (a.attached_to))) {
+          if (this._isMounted) {
+            this.setState({ addonAttachments, attachmentsLoaded: true });
+          }
+        }
+      });
+    });
+  }
+
+  loadAddons() {
+    Promise.all([
+      api.getAppAddons(this.props.app),
+      api.getAddonAttachments(this.props.app),
+    ]).then(([r1, r2]) => {
+      if (this._isMounted) {
+        this.setState({
+          addons: r1.data,
+          addonAttachments: r2.data,
+          loading: false,
+        });
+        this.getAppsAttachedToAddon();
+      }
+    });
+  }
+
+  formatAttachment(attachment, index) { // eslint-disable-line class-methods-use-this
+    return (
+      <TableRow className={`attachment-${index}`} style={style.tableRow} key={attachment.id}>
+        <TableCell colSpan="2">
+          <div className="attachment-name" style={style.tableRowColumn.title}>{attachment.name}</div>
+          <div style={style.tableRowColumn.sub}>{attachment.id}</div>
+        </TableCell>
+        <TableCell>
+          {attachment.owner && (
+            <div className="attachment-owner" style={{ color: muiTheme.palette.secondary.main }}>Owner</div>
+          )}
+        </TableCell>
+      </TableRow>
+    );
+  }
+
+  handleAddonDialogClose = () => {
+    this.setState({ addonDialogOpen: false });
   }
 
   handleNewAddon = () => {
@@ -108,6 +269,14 @@ export default class Addons extends Component {
 
   handleNewAddonCancel = () => {
     this.setState({ new: false });
+  }
+
+  handleAttachAddon = () => {
+    this.setState({ attach: true });
+  }
+
+  handleAttachAddonCancel = () => {
+    this.setState({ attach: false });
   }
 
   handleRemoveAddon = () => {
@@ -120,22 +289,55 @@ export default class Addons extends Component {
         submitFail: true,
         loading: false,
         new: false,
-        confirmOpen: false,
+        confirmAddonOpen: false,
+        confirmAttachmentOpen: false,
+        attach: false,
       });
     });
   }
 
-  handleConfirmation = (addon) => {
+  handleRemoveAddonAttachment = () => {
+    this.setState({ loading: true });
+    api.deleteAddonAttachment(this.props.app, this.state.attachment.id).then(() => {
+      this.reload('Attachment Deleted');
+    }).catch((error) => {
+      this.setState({
+        submitMessage: error.response.data,
+        submitFail: true,
+        loading: false,
+        new: false,
+        confirmAddonOpen: false,
+        confirmAttachmentOpen: false,
+        attach: false,
+      });
+    });
+  }
+
+  handleAddonConfirmation = (addon) => {
     this.setState({
-      confirmOpen: true,
+      confirmAddonOpen: true,
       addon,
     });
   }
 
-  handleCancelConfirmation = () => {
+  handleCancelAddonConfirmation = () => {
     this.setState({
-      confirmOpen: false,
+      confirmAddonOpen: false,
       addon: null,
+    });
+  }
+
+  handleAddonAttachmentConfirmation = (attachment) => {
+    this.setState({
+      confirmAttachmentOpen: true,
+      attachment,
+    });
+  }
+
+  handleCancelAddonAttachmentConfirmation = () => {
+    this.setState({
+      confirmAttachmentOpen: false,
+      attachment: null,
     });
   }
 
@@ -149,82 +351,174 @@ export default class Addons extends Component {
 
   reload = (message) => {
     this.setState({ loading: true });
-    api.getAppAddons(this.props.app).then((response) => {
+    Promise.all([
+      api.getAppAddons(this.props.app),
+      api.getAddonAttachments(this.props.app),
+    ]).then(([r1, r2]) => {
       this.setState({
-        addons: response.data,
+        addons: r1.data,
+        addonAttachments: r2.data,
         loading: false,
         new: false,
         message,
         open: true,
-        confirmOpen: false,
+        confirmAddonOpen: false,
+        confirmAttachmentOpen: false,
+        attach: false,
       });
+      this.getAppsAttachedToAddon();
     });
   }
 
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
           </div>
         </MuiThemeProvider>
       );
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
-        <div>
-          {!this.state.new && (
-            <Paper zDepth={0}>
-              <IconButton
-                className="new-addon"
-                onTouchTap={this.handleNewAddon}
-                tooltip="New Addon"
-                tooltipPosition="bottom-left"
-              >
-                <AddIcon />
-              </IconButton>
-            </Paper>
+      <MuiThemeProvider theme={muiTheme}>
+        <div style={{ overflow: 'visible' }}>
+          {(!this.state.new && !this.state.attach) && (
+            <div>
+              <Tooltip title="New Addon" placement="bottom-end">
+                <IconButton
+                  className="new-addon"
+                  onClick={this.handleNewAddon}
+                  style={style.iconButton}
+                >
+                  <AddIcon />
+                </IconButton>
+              </Tooltip>
+              <Tooltip title="Attach Addon" placement="bottom-end">
+                <IconButton
+                  className="attach-addon"
+                  onClick={this.handleAttachAddon}
+                  style={style.iconButton}
+                >
+                  <AttachIcon />
+                </IconButton>
+              </Tooltip>
+            </div>
           )}
           {this.state.new && (
             <div>
-              <IconButton className="addon-cancel" onTouchTap={this.handleNewAddonCancel}><RemoveIcon /></IconButton>
+              <IconButton style={style.iconButton} className="addon-cancel" onClick={this.handleNewAddonCancel}><RemoveIcon /></IconButton>
               <NewAddon app={this.props.app} onComplete={this.reload} />
             </div>
           )}
+          {this.state.attach && (
+            <div>
+              <IconButton style={style.iconButton} className="attach-cancel" onClick={this.handleAttachAddonCancel}><RemoveIcon /></IconButton>
+              <AttachAddon app={this.props.app} onComplete={this.reload} />
+            </div>
+          )}
           <Table className="addon-list">
-            <TableHeader adjustForCheckbox={false} displaySelectAll={false} selectable={false}>
+            <TableHead>
               <TableRow>
-                <TableHeaderColumn>Addon</TableHeaderColumn>
-                <TableHeaderColumn>Plan</TableHeaderColumn>
-                <TableHeaderColumn style={style.tableRowColumn.icon}>Remove</TableHeaderColumn>
+                <TableCell>Addon</TableCell>
+                <TableCell>Plan</TableCell>
+                <TableCell style={style.tableRowColumn.icon}>Remove</TableCell>
               </TableRow>
-            </TableHeader>
-            <TableBody displayRowCheckbox={false} showRowHover selectable={false}>
-              {this.getAddons()}
+            </TableHead>
+            <TableBody>
+              {this.state.addons.length > 0 ? this.getAddons() : (
+                <TableRow><TableCell /><TableCell>No Results</TableCell><TableCell /></TableRow>
+              )}
             </TableBody>
           </Table>
-          <ConfirmationModal className="remove-confirm" open={this.state.confirmOpen} onOk={this.handleRemoveAddon} onCancel={this.handleCancelConfirmation} message="Are you sure you want to delete this addon?" />
+          {this.state.addonAttachments.length > 0 && (
+            <Table className="addon-attachment-list">
+              <TableHead>
+                <TableRow>
+                  <TableCell>Attachment</TableCell>
+                  <TableCell>Plan</TableCell>
+                  <TableCell>Source</TableCell>
+                  <TableCell style={style.tableRowColumn.icon}>Remove</TableCell>
+                </TableRow>
+              </TableHead>
+              <TableBody>
+                {this.getAddonAttachments()}
+              </TableBody>
+            </Table>
+          )}
+          <Dialog
+            className="attached-apps-dialog"
+            onClose={this.handleAddonDialogClose}
+            open={this.state.addonDialogOpen}
+          >
+            <DialogTitle>
+              {this.getDialogTitle()}
+            </DialogTitle>
+            <DialogContent>
+              {!isEmpty(this.state.currentAddon) && (
+                <Table>
+                  <TableHead>
+                    <TableRow>
+                      <TableCell colSpan="2">App</TableCell>
+                      <TableCell>Ownership</TableCell>
+                    </TableRow>
+                  </TableHead>
+                  <TableBody>
+                    {this.state.currentAddon.attached_to &&
+                      this.state.currentAddon.attached_to.map((attachment, index) =>
+                        this.formatAttachment(attachment, index),
+                      )}
+                  </TableBody>
+                </Table>
+              )}
+            </DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={this.handleAddonDialogClose}
+              >
+                Ok
+              </Button>
+            </DialogActions>
+          </Dialog>
+          <ConfirmationModal
+            className="remove-addon-confirm"
+            open={this.state.confirmAddonOpen}
+            onOk={this.handleRemoveAddon}
+            onCancel={this.handleCancelAddonConfirmation}
+            message="Are you sure you want to delete this addon?"
+          />
+          <ConfirmationModal
+            className="remove-attachment-confirm"
+            open={this.state.confirmAttachmentOpen}
+            onOk={this.handleRemoveAddonAttachment}
+            onCancel={this.handleCancelAddonAttachmentConfirmation}
+            message="Are you sure you want to delete this attachment?"
+          />
           <Dialog
             className="addon-error"
             open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleDialogClose}
-              />}
           >
-            {this.state.submitMessage}
+            <DialogContent>
+              <DialogContentText>{this.state.submitMessage}</DialogContentText>
+            </DialogContent>
+            <DialogActions>
+              <Button
+                className="ok"
+                color="primary"
+                onClick={this.handleDialogClose}
+              >
+                Ok
+              </Button>
+            </DialogActions>
           </Dialog>
           <Snackbar
             className="addon-snack"
             open={this.state.open}
             message={this.state.message}
             autoHideDuration={3000}
-            onRequestClose={this.handleRequestClose}
+            onClose={this.handleRequestClose}
           />
         </div>
       </MuiThemeProvider>
@@ -234,6 +528,4 @@ export default class Addons extends Component {
 
 Addons.propTypes = {
   app: PropTypes.string.isRequired,
-  active: PropTypes.bool.isRequired,
 };
-
