@@ -1,13 +1,25 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Table, TableBody, TableRow, TableRowColumn, TableHeader, TableHeaderColumn } from 'material-ui/Table';
+import {
+  Table, TableBody, TableHead, TableRow, TableCell, TableFooter, TablePagination,
+} from '@material-ui/core';
+import { createMuiTheme, MuiThemeProvider } from '@material-ui/core/styles';
+
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+});
 
 const style = {
   tableRow: {
     height: '58px',
     cursor: 'pointer',
   },
-  tableRowColumn: {
+  tableCell: {
     title: {
       fontSize: '16px',
     },
@@ -19,46 +31,87 @@ const style = {
 };
 
 export default class SitesList extends Component {
-  getSites() {
-    return this.props.sites.map((site) => {
+  constructor(props, context) {
+    super(props, context);
+    this.state = {
+      page: 0,
+      rowsPerPage: 15,
+    };
+  }
+
+  getSites(page, rowsPerPage) {
+    return this.props.sites.slice(page * rowsPerPage, (page * rowsPerPage) + rowsPerPage).map((site) => {
       const date = new Date(site.updated_at);
       return (
-        <TableRow className={site.domain} key={site.id} style={style.tableRow}>
-          <TableRowColumn>
-            <div style={style.tableRowColumn.title}>{site.domain}</div>
-            <div style={style.tableRowColumn.sub}>{site.id}</div>
-          </TableRowColumn>
-          <TableRowColumn>
+        <TableRow
+          className={site.domain}
+          key={site.id}
+          style={style.tableRow}
+          hover
+          onClick={() => this.handleRowSelection(site.id)}
+        >
+          <TableCell>
+            <div style={style.tableCell.title}>{site.domain}</div>
+            <div style={style.tableCell.sub}>{site.id}</div>
+          </TableCell>
+          <TableCell>
             <div>{date.toLocaleString()}</div>
-          </TableRowColumn>
-          <TableRowColumn>
-            <div style={style.tableRowColumn.title}>{site.region.name}</div>
-          </TableRowColumn>
+          </TableCell>
+          <TableCell>
+            <div style={style.tableCell.title}>{site.region.name}</div>
+          </TableCell>
         </TableRow>
       );
     });
   }
 
-  handleRowSelection = (selectedRows) => {
-    window.location = `#/sites/${this.props.sites[selectedRows].domain}`;
+  handleRowSelection = (id) => {
+    window.location = `#/sites/${id}/info`;
   }
 
+  handleChangePage = (event, page) => {
+    this.setState({ page });
+  };
+
+  handleChangeRowsPerPage = (event) => {
+    this.setState({ rowsPerPage: event.target.value });
+  };
+
   render() {
+    const { sites } = this.props;
+    const { page, rowsPerPage } = this.state;
     return (
-      <div>
-        <Table className="site-list" onRowSelection={this.handleRowSelection} wrapperStyle={{ overflow: 'visible' }} bodyStyle={{ overflow: 'visible' }}>
-          <TableHeader adjustForCheckbox={false} displaySelectAll={false} selectable={false}>
-            <TableRow>
-              <TableHeaderColumn>Site</TableHeaderColumn>
-              <TableHeaderColumn>Updated</TableHeaderColumn>
-              <TableHeaderColumn>Region</TableHeaderColumn>
-            </TableRow>
-          </TableHeader>
-          <TableBody displayRowCheckbox={false} showRowHover>
-            {this.getSites()}
-          </TableBody>
-        </Table>
-      </div>
+      <MuiThemeProvider theme={muiTheme}>
+        <div>
+          <Table className="site-list">
+            <TableHead>
+              <TableRow>
+                <TableCell>Site</TableCell>
+                <TableCell>Updated</TableCell>
+                <TableCell>Region</TableCell>
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              {this.getSites(page, rowsPerPage)}
+            </TableBody>
+            {sites.length !== 0 && (
+              <TableFooter>
+                <TableRow>
+                  <TablePagination
+                    rowsPerPageOptions={[15, 25, 50]}
+                    colSpan={3}
+                    count={sites.length}
+                    rowsPerPage={rowsPerPage}
+                    page={page}
+                    onChangePage={this.handleChangePage}
+                    onChangeRowsPerPage={this.handleChangeRowsPerPage}
+                  />
+                </TableRow>
+              </TableFooter>
+            )}
+          </Table>
+        </div>
+      </MuiThemeProvider>
     );
   }
 }
