@@ -1,54 +1,95 @@
 import React, { Component } from 'react';
-import getMuiTheme from 'material-ui/styles/getMuiTheme';
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import {
+  Button, IconButton, Dialog, DialogContent, DialogTitle, DialogContentText, DialogActions,
+  Tab, Tabs, CircularProgress, Snackbar, Card, CardHeader,
+} from '@material-ui/core';
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
 import PropTypes from 'prop-types';
-import { Card, CardHeader } from 'material-ui/Card';
-import FlatButton from 'material-ui/FlatButton';
-import { Tabs, Tab } from 'material-ui/Tabs';
-import RefreshIndicator from 'material-ui/RefreshIndicator';
-import Dialog from 'material-ui/Dialog';
-import Snackbar from 'material-ui/Snackbar';
-import LaptopIcon from 'material-ui/svg-icons/hardware/computer';
-import GlobeIcon from 'material-ui/svg-icons/social/public';
-import Forward from 'material-ui/svg-icons/navigation/arrow-forward';
-import RemoveIcon from 'material-ui/svg-icons/content/clear';
-import IconButton from 'material-ui/IconButton';
+
+import LaptopIcon from '@material-ui/icons/Computer';
+import GlobeIcon from '@material-ui/icons/Public';
+import Forward from '@material-ui/icons/ArrowForward';
+import RemoveIcon from '@material-ui/icons/Clear';
 
 import api from '../../services/api';
 import ConfirmationModal from '../../components/ConfirmationModal';
 import { Stage } from '../../components/Pipelines';
 
-const muiTheme = getMuiTheme({
-  fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
-  tabs: {
-    backgroundColor: '#3c4146',
+const muiTheme = createMuiTheme({
+  palette: {
+    primary: { main: '#0097a7' },
+  },
+  typography: {
+    fontFamily: '-apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"',
+  },
+  overrides: {
+    MuiTabs: {
+      root: {
+        backgroundColor: '#3c4146',
+        color: 'white',
+        maxWidth: '1024px',
+      },
+    },
+    MuiTab: {
+      root: {
+        minWidth: '120px !important',
+      },
+    },
+    MuiCardContent: {
+      root: {
+        display: 'flex',
+        flexFlow: 'row-reverse',
+        padding: '0px 16px 0px 0px !important',
+      },
+    },
+    MuiCard: {
+      root: {
+        maxWidth: '1024px',
+        marginLeft: 'auto',
+        marginRight: 'auto',
+        marginTop: '12px',
+      },
+    },
+    MuiCardHeader: {
+      title: {
+        fontSize: '15px',
+        fontWeight: '500',
+      },
+      subheader: {
+        fontSize: '14px',
+        fontWeight: '500',
+      },
+    },
   },
 });
 
 const style = {
+  iconButton: {
+    color: 'black',
+    float: 'right',
+  },
   refresh: {
     div: {
       marginLeft: 'auto',
       marginRight: 'auto',
       width: '40px',
       height: '40px',
-      marginTop: '20%',
+      marginTop: '15%',
+      marginBottom: '5%',
     },
     indicator: {
       display: 'inline-block',
       position: 'relative',
+      color: 'white',
     },
   },
   card: {
-    maxWidth: '1024px',
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    marginTop: '12px',
-  },
-  iconButton: {
-    float: 'right',
+    overflow: 'visible',
+    marginBottom: '20px',
   },
 };
+
+const tabs = ['review', 'development', 'staging', 'production'];
 
 export default class PipelineInfo extends Component {
   constructor(props) {
@@ -61,15 +102,21 @@ export default class PipelineInfo extends Component {
       confirmOpen: false,
       message: '',
       pipeline: null,
-      reviewActive: true,
-      devActive: false,
-      stagingActive: false,
-      prodActive: false,
+      currentTab: 'review',
+      baseHash: `#/pipelines/${this.props.match.params.pipeline}/`,
+      basePath: `/pipelines/${this.props.match.params.pipeline}`,
     };
   }
 
   componentDidMount() {
     api.getPipeline(this.props.match.params.pipeline).then((response) => {
+      const hashPath = window.location.hash;
+      let currentTab = hashPath.replace(this.state.baseHash, '');
+      if (!tabs.includes(currentTab)) {
+        currentTab = 'review';
+        window.location.hash = `${this.state.baseHash}review`;
+      }
+      this.setState({ currentTab });
       this.setState({
         pipeline: response.data,
         loading: false,
@@ -80,6 +127,31 @@ export default class PipelineInfo extends Component {
         submitMessage: error.response.data,
       });
     });
+  }
+
+  componentDidUpdate(prevProps) {
+    // If we changed locations AND it was a 'pop' history event (back or forward button)
+    const routeHasChanged = prevProps.location.pathname !== this.props.location.pathname;
+    if (routeHasChanged && this.props.history.action === 'POP') {
+      // If hitting back took us to the base path without a tab, hit back again
+      // TODO: what if we hit forward to the base path?
+      //    detect forward click and do window.history.forward()
+      if (this.props.location.pathname === `${this.state.basePath}` ||
+          this.props.location.pathname === `${this.state.basePath}/`) {
+        window.history.back();
+        return;
+      }
+      const hashPath = window.location.hash;
+      if (hashPath.includes(this.state.baseHash)) {
+        let currentTab = hashPath.replace(this.state.baseHash, '');
+        if (!tabs.includes(currentTab)) {
+          currentTab = 'review';
+          window.location = `${this.state.baseHash}review`;
+        }
+        // Since we check conditions before setState we avoid infinite loops
+        this.setState({ currentTab }); // eslint-disable-line react/no-did-update-set-state
+      }
+    }
   }
 
   handleRemovePipeline = () => {
@@ -130,129 +202,166 @@ export default class PipelineInfo extends Component {
     });
   }
 
-  reviewTabActive = () => {
-    this.setState({
-      reviewActive: true,
-      devActive: false,
-      stagingActive: false,
-      prodActive: false,
-    });
+  changeActiveTab = (event, newTab) => {
+    if (this.state.currentTab !== newTab) {
+      this.setState({
+        currentTab: newTab,
+      });
+      this.props.history.push(`${newTab}`);
+    }
   }
-  devTabActive = () => {
-    this.setState({
-      reviewActive: false,
-      devActive: true,
-      stagingActive: false,
-      prodActive: false,
-    });
-  }
-  stagingTabActive = () => {
-    this.setState({
-      reviewActive: false,
-      devActive: false,
-      stagingActive: true,
-      prodActive: false,
-    });
-  }
-  prodTabActive = () => {
-    this.setState({
-      reviewActive: false,
-      devActive: false,
-      stagingActive: false,
-      prodActive: true,
-    });
+
+  renderTabContent = () => {
+    const { currentTab } = this.state;
+    return (
+      <React.Fragment>
+        {currentTab === 'review' && (
+          <Stage
+            stage="review"
+            pipeline={this.state.pipeline}
+            onError={this.handleError}
+            onAlert={this.handleAlert}
+            active={this.state.currentTab === 'review'}
+          />
+        )}
+        {currentTab === 'development' && (
+          <Stage
+            stage="development"
+            pipeline={this.state.pipeline}
+            onError={this.handleError}
+            onAlert={this.handleAlert}
+            active={this.state.currentTab === 'development'}
+          />
+        )}
+        {currentTab === 'staging' && (
+          <Stage
+            stage="staging"
+            pipeline={this.state.pipeline}
+            onError={this.handleError}
+            onAlert={this.handleAlert}
+            active={this.state.currentTab === 'staging'}
+          />
+        )}
+        {currentTab === 'production' && (
+          <Stage
+            stage="production"
+            pipeline={this.state.pipeline}
+            onError={this.handleError}
+            onAlert={this.handleAlert}
+            active={this.state.currentTab === 'production'}
+          />
+        )}
+      </React.Fragment>
+    );
   }
 
   render() {
     if (this.state.loading) {
       return (
-        <MuiThemeProvider muiTheme={muiTheme}>
+        <MuiThemeProvider theme={muiTheme}>
           <div style={style.refresh.div}>
-            <RefreshIndicator top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
+            <CircularProgress top={0} size={40} left={0} style={style.refresh.indicator} status="loading" />
             <Dialog
               className="not-found-error"
               open={this.state.submitFail}
-              modal
-              actions={
-                <FlatButton
-                  className="ok"
-                  label="Ok"
-                  primary
-                  onTouchTap={this.handleNotFoundClose}
-                />}
             >
-              {this.state.submitMessage}
+              <DialogTitle>Error</DialogTitle>
+              <DialogContent>
+                <DialogContentText>{this.state.submitMessage}</DialogContentText>
+              </DialogContent>
+              <DialogActions>
+                <Button
+                  onClick={this.handleNotFoundClose}
+                  color="primary"
+                >
+                  Ok
+                </Button>
+              </DialogActions>
             </Dialog>
           </div>
         </MuiThemeProvider>);
     }
     return (
-      <MuiThemeProvider muiTheme={muiTheme}>
+      <MuiThemeProvider theme={muiTheme}>
         <div>
           <Card className="card" style={style.card}>
             <CardHeader
               className="header"
               title={this.state.pipeline.name}
-              subtitle={this.state.pipeline.id}
+              subheader={this.state.pipeline.id}
+              action={
+                <IconButton className="delete-pipeline" style={style.iconButton} onClick={this.handleConfirmation}>
+                  <RemoveIcon />
+                </IconButton>
+              }
+            />
+            <Tabs
+              fullWidth
+              value={this.state.currentTab}
+              onChange={this.changeActiveTab}
+              scrollButtons="off"
             >
-              <IconButton className="delete-pipeline" style={style.iconButton} onTouchTap={this.handleConfirmation}><RemoveIcon /></IconButton>
-              <ConfirmationModal open={this.state.confirmOpen} onOk={this.handleRemovePipeline} onCancel={this.handleCancelConfirmation} message="Are you sure you want to delete this pipeline?" />
-            </CardHeader>
-            <Tabs >
               <Tab
+                disableRipple
                 className="review-tab"
                 icon={<LaptopIcon />}
                 label="Review"
-                onActive={this.reviewTabActive}
-              >
-                <Stage stage="review" pipeline={this.state.pipeline} onError={this.handleError} onAlert={this.handleAlert} active={this.state.reviewActive} />
-              </Tab>
+                value="review"
+              />
               <Tab
+                disableRipple
                 className="dev-tab"
                 icon={<Forward />}
                 label="Development"
-                onActive={this.devTabActive}
-              >
-                <Stage stage="development" pipeline={this.state.pipeline} onError={this.handleError} onAlert={this.handleAlert} active={this.state.devActive} />
-              </Tab>
+                value="development"
+              />
               <Tab
+                disableRipple
                 className="staging-tab"
                 icon={<Forward />}
                 label="Staging"
-                onActive={this.stagingTabActive}
-              >
-                <Stage stage="staging" pipeline={this.state.pipeline} onError={this.handleError} onAlert={this.handleAlert} active={this.state.stagingActive} />
-              </Tab>
+                value="staging"
+              />
               <Tab
+                disableRipple
                 className="prod-tab"
                 icon={<GlobeIcon />}
                 label="Production"
-                onActive={this.prodTabActive}
-              >
-                <Stage stage="production" pipeline={this.state.pipeline} onError={this.handleError} onAlert={this.handleAlert} active={this.state.prodActive} />
-              </Tab>
+                value="production"
+              />
             </Tabs>
+            {this.renderTabContent()}
           </Card>
+          <ConfirmationModal
+            open={this.state.confirmOpen}
+            onOk={this.handleRemovePipeline}
+            onCancel={this.handleCancelConfirmation}
+            message="Are you sure you want to delete this pipeline?"
+          />
           <Dialog
             className="error"
             open={this.state.submitFail}
-            modal
-            actions={
-              <FlatButton
-                className="ok"
-                label="Ok"
-                primary
-                onTouchTap={this.handleClose}
-              />}
           >
-            {this.state.submitMessage}
+            <DialogTitle>Error</DialogTitle>
+            <DialogContent>
+              <DialogContentText>{this.state.submitMessage}</DialogContentText>
+            </DialogContent>
+            <DialogActions>
+              <Button
+                color="primary"
+                onClick={this.handleClose}
+                className="ok"
+              >
+              Ok
+              </Button>
+            </DialogActions>
           </Dialog>
           <Snackbar
             className="pipeline-snack"
             open={this.state.open}
             message={this.state.message}
             autoHideDuration={3000}
-            onRequestClose={this.handleRequestClose}
+            onClose={this.handleRequestClose}
           />
         </div>
       </MuiThemeProvider>
@@ -262,4 +371,6 @@ export default class PipelineInfo extends Component {
 
 PipelineInfo.propTypes = {
   match: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  history: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
+  location: PropTypes.object.isRequired, // eslint-disable-line react/forbid-prop-types
 };
